requirements.txt
appdirs==1.4.4
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.0
charset-normalizer==2.0.4
cmake==3.21.0
cycler==0.10.0
Cython==0.29.24
dataclasses==0.6
decorator==5.0.9
distlib==0.3.2
filelock==3.0.12
filterpy==1.4.5
flatbuffers==2.0
future==0.18.2
idna==3.2
imageio==2.9.0
importlib-metadata==4.6.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.30.0
matplotlib==3.3.4
matplotlib-inline==0.1.2
mkl-fft==1.0.6
mkl-random==1.0.1.1
numba==0.50.0
numpy==1.18
olefile==0.46
opencv-python==4.1.2.30
pandas==1.1.5
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.17.2
scikit-learn==0.22.2
scipy==1.5.4
seaborn==0.11.2
six
sklearn==0.0
stevedore==3.3.0
threadpoolctl==2.2.0
torch==1.8.1
torchvision==0.9.1
tqdm
typing==3.7.4.3
typing-extensions
urllib3==1.26.6
wcwidth==0.2.5
zipp==3.4.1